fix(skills): pin ratio casts to DOUBLE so SQLite and DuckDB agree (#371)
SQLite REAL is float64; DuckDB REAL is float32. Skill SQL owns the
width: CAST(... AS DOUBLE) before ROUND so derived percentages and
rates are identical across backends, including when the parquet cache
is absent. Python round() would only paper over the exported values.

+++ src/nsys_ai/skills/builtins/cpu_gpu_pipeline.py
@@ -56,7 +56,7 @@ SELECT
     max_queue_delay_us,
     starvation_events,
     avg_kernel_us,
-    ROUND(CAST(dispatches AS REAL) / (SELECT SUM(dispatches) FROM per_thread) * 100, 1)
+    ROUND(CAST(dispatches AS DOUBLE) / (SELECT SUM(dispatches) FROM per_thread) * 100, 1)
         AS pct_of_dispatches
 FROM per_thread
 ORDER BY dispatches DESC, cpu_tid ASC
